requirements.txt
pytest==5.2.2
py
requests
testcontainers
selenium
psutil
pytest-html
allure-pytest
sqlalchemy
pymysql

